+++ example/a.ts
@@ -1,9 +1,34 @@
-import { Elysia } from '../src'
+import { Elysia, t } from '../src'
 
-const app = new Elysia().get('/', function* ({ status }) {
-	const project = { url: '1234' }
-	if (!project) return status(400)
-	return status(200, project.url)
+const MessageSchema = t.Object({
+	message: t.String()
 })
 
-app['~Routes'].get.response
+const app = new Elysia({
+	name: 'appService'
+})
+	.model({
+		message: MessageSchema
+	})
+	.get(
+		'/health',
+		() => {
+			return {
+				timestamp: new Date().toISOString(),
+				status: 'ok'
+			}
+		},
+		{
+			response: {
+				200: t.Object({
+					status: t.String(),
+					timestamp: t.String()
+				}),
+				default: t.Ref('message')
+			}
+		}
+	)
+
+app.handle(new Request('http://localhost/health'))
+	.then((x) => x.json())
+	.then(console.log)