+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# 1.3.2 - 24 May 2025
+# 1.3.2 - 27 May 2025
 Feature:
 - Support Bun native static response per method for Bun >= 1.2.14
 - [#1213](https://github.com/elysiajs/elysia/issues/1213) trace.time is undefined in .trace() callback
@@ -7,6 +7,7 @@ Improvement:
 - implement all universal type
 - offload `AsyncGenerator`, `ReplaceFile` from Eden Treaty to `CreateEden`
 - [#1223](https://github.com/elysiajs/elysia/issues/1223) infer `status(200)` response from handler if not specified
+- [#1185](https://github.com/elysiajs/elysia/issues/1185) use non-greedy match for `isContextPassToFunction` to prevent false positive
 
 # 1.3.1 - 8 May 2025
 Bug fix: